+++ bin/json.dart
@@ -2,6 +2,7 @@
 import 'common.dart';
 import 'params.dart';
 import 'aux.dart';
+import 'riswhois.dart' show info2map;
 
 Map<String, dynamic> getMappedHops(List<Hop> stat, int hops, String host) {
   List<Map<String, dynamic>> all = [];
@@ -22,11 +23,18 @@ Map<String, dynamic> _hop2map(Hop h, int ttl) {
     'avg': _todbl(h.avg), 'jttr': _todbl(h.jttr)
   };
   tm.removeWhere((k, v) => (v is String) ? v.isEmpty : false);
-  var all = {
-    'ttl': ttl, 'addr': h.addr, 'name': h.name,
-    'sent': h.data.sent, 'rcvd': h.data.rcvd, 'loss': h.loss,
-  };
-  all.removeWhere((k, v) => (v is List) ? v.isEmpty : false);
+  List<Map<String, dynamic>> hostinfo = [];
+  for (var i in h.info) {
+    Map<String, dynamic> m = {};
+    if (i.addr != null) m['addr'] = i.addr;
+    if (i.name != null) m['name'] = i.name;
+    var w = i.whois;
+    if (w != null) m.addAll(info2map(w));
+    if (m.isNotEmpty) hostinfo.add(m);
+  }
+  //
+  var all = {'ttl': ttl, 'sent': h.data.sent, 'rcvd': h.data.rcvd, 'loss': h.loss};
+  if (hostinfo.isNotEmpty) all['host'] = hostinfo;
   if (tm.isNotEmpty) all.addAll({'timeunit': 'millisecond', 'timing': tm});
   if (h.unreach) all[_extra] = unreachMesg;
   var mesg = h.wrong;